requirements.txt
PySide6>=6.8
numpy>=2.0
PyYAML>=6.0
pyserial>=3.5
sounddevice>=0.5
pytest>=8.0
